fakeweb/requirements.txt
flask
python-whois
requests
dnspython